Add checks for Kerberos properties in Hive connector
Ensure that we have all the properties set for Hive connector
when using Kerberos to access Hive metastore. Extracted Hive
Kerberos properties out of HiveClientConfig to add not null checks.
