+++ presto-hive/src/main/java/com/facebook/presto/hive/authentication/AuthenticationModules.java
@@ -15,7 +15,7 @@ package com.facebook.presto.hive.authentication;
 
 import com.facebook.presto.hive.ForHdfs;
 import com.facebook.presto.hive.ForHiveMetastore;
-import com.facebook.presto.hive.HiveClientConfig;
+import com.facebook.presto.hive.HiveKerberosConfig;
 import com.google.inject.Binder;
 import com.google.inject.Key;
 import com.google.inject.Module;
@@ -25,6 +25,7 @@ import com.google.inject.Singleton;
 import javax.inject.Inject;
 
 import static com.google.inject.Scopes.SINGLETON;
+import static io.airlift.configuration.ConfigBinder.configBinder;
 
 public final class AuthenticationModules
 {
@@ -48,15 +49,16 @@ public final class AuthenticationModules
                 binder.bind(HiveMetastoreAuthentication.class)
                         .to(KerberosHiveMetastoreAuthentication.class)
                         .in(SINGLETON);
+                configBinder(binder).bindConfig(HiveKerberosConfig.class);
             }
 
             @Provides
             @Singleton
             @ForHiveMetastore
-            HadoopAuthentication createHadoopAuthentication(HiveClientConfig hiveClientConfig)
+            HadoopAuthentication createHadoopAuthentication(HiveKerberosConfig config)
             {
-                String principal = hiveClientConfig.getHiveMetastoreClientPrincipal();
-                String keytabLocation = hiveClientConfig.getHiveMetastoreClientKeytab();
+                String principal = config.getHiveMetastoreClientPrincipal();
+                String keytabLocation = config.getHiveMetastoreClientKeytab();
                 return createCachingKerberosHadoopAuthentication(principal, keytabLocation);
             }
         };
@@ -91,16 +93,17 @@ public final class AuthenticationModules
                 binder.bind(HdfsAuthentication.class)
                         .to(DirectHdfsAuthentication.class)
                         .in(SINGLETON);
+                configBinder(binder).bindConfig(HiveKerberosConfig.class);
             }
 
             @Inject
             @Provides
             @Singleton
             @ForHdfs
-            HadoopAuthentication createHadoopAuthentication(HiveClientConfig hiveClientConfig)
+            HadoopAuthentication createHadoopAuthentication(HiveKerberosConfig config)
             {
-                String principal = hiveClientConfig.getHdfsPrestoPrincipal();
-                String keytabLocation = hiveClientConfig.getHdfsPrestoKeytab();
+                String principal = config.getHdfsPrestoPrincipal();
+                String keytabLocation = config.getHdfsPrestoKeytab();
                 return createCachingKerberosHadoopAuthentication(principal, keytabLocation);
             }
         };
@@ -116,16 +119,17 @@ public final class AuthenticationModules
                 binder.bind(HdfsAuthentication.class)
                         .to(ImpersonatingHdfsAuthentication.class)
                         .in(SINGLETON);
+                configBinder(binder).bindConfig(HiveKerberosConfig.class);
             }
 
             @Inject
             @Provides
             @Singleton
             @ForHdfs
-            HadoopAuthentication createHadoopAuthentication(HiveClientConfig hiveClientConfig)
+            HadoopAuthentication createHadoopAuthentication(HiveKerberosConfig config)
             {
-                String principal = hiveClientConfig.getHdfsPrestoPrincipal();
-                String keytabLocation = hiveClientConfig.getHdfsPrestoKeytab();
+                String principal = config.getHdfsPrestoPrincipal();
+                String keytabLocation = config.getHdfsPrestoKeytab();
                 return createCachingKerberosHadoopAuthentication(principal, keytabLocation);
             }
         };
